useQuery를 사용하여 comments를 불러와라 - xtring
