+++ base-blog-em/src/PostDetail.jsx
@@ -1,41 +1,13 @@
-async function fetchComments(postId) {
-  const response = await fetch(
-    `https://jsonplaceholder.typicode.com/comments?postId=${postId}`
-  );
-  return response.json();
-}
-
-async function deletePost(postId) {
-  const response = await fetch(
-    `https://jsonplaceholder.typicode.com/postId/${postId}`,
-    { method: "DELETE" }
-  );
-  return response.json();
-}
-
-async function updatePost(postId) {
-  const response = await fetch(
-    `https://jsonplaceholder.typicode.com/postId/${postId}`,
-    { method: "PATCH", data: { title: "REACT QUERY FOREVER!!!!" } }
-  );
-  return response.json();
-}
+import PostComments from "./PostComments";
 
 export function PostDetail({ post }) {
-  // replace with useQuery
-  const data = [];
-
   return (
     <>
       <h3 style={{ color: "blue" }}>{post.title}</h3>
       <button>Delete</button> <button>Update title</button>
       <p>{post.body}</p>
       <h4>Comments</h4>
-      {data.map((comment) => (
-        <li key={comment.id}>
-          {comment.email}: {comment.body}
-        </li>
-      ))}
+      <PostComments postId={post.id} />
     </>
   );
 }
